Move `OutputPartBuffer` to output package
